Changed purpose of limelight system
A limelight system is now used for an individual pipeline rather than for an individual limelight.  This enables things such as the following:
```
private Limelight cubeCamera = new Limelight(1); // The x position of a cube can now be determined with the double cubeCamera.X()
private Limelight coneCamera = new Limelight(5); // The x position of a cone can now be determined separately from that of the cube, with the simple double cubeCamera.X()
```
With this method, using limelight may be easier than ever, especially when dealing with different types of objects.