Added Better Direction Functionality
Direction now starts at 180 degrees, as the robot will be immediately scoring at the beginning of autonomous.  Also, we have experienced frequent issues with the navx getting off course over time.  We find that these issues are solved when we press the reset button on the navx.  Now, by pressing X on the secondary controller, we are able to perform this same action in case of an emergency.