Use width factor for pickup closeness
Previously, we were using the area of the visible object to determine how far away the object was.  This turned out to have some disastrous results: because our camera is facing up at an angle, the lower part of the object goes out of view as we near it.  This makes the area DECREASE as we go toward it - a situation the programming is not made to encounter.  Now, we use the WIDTH of the object rather than its area, so we can get an accurate representation of the objects distance from the camera.