Add `drop` Function with Data Tracking
When this function is run to drop an object, variables adjust to keep track of the following data and post it to the dash:
 - Cubes Scored
 - Cones Scored
 - Points Earned
This code does account for the different amounts of points earned in autonomous vs teleop.  A game piece scored in autonomous is worth one point more than one scored in teleop.