Reduce zeroing speed to not kill claw when zeroing
The code before would have been much too powerful for zeroing, especially considering the mechanical advantage of the acme rod.  Zeroing makes the claw close until it detects that it can't close any more, so it knows that it is at the zero position.  With speed 0.5, the claw would have had so much power that it would not have stopped when it was all the way closed; it would have destroyed the claw and kept spinning.  With speed 0.1, there will not be enough power to do this, and the claw will come to a stop and know where it is at.