Keep good changes from bad pull request
These are simple changes to the code that shouldn't affect its functionality too much.  Because pull request #12, started by issue #6, many changes were made to the programming.  As described in the last comment in the #12 pull request chat, some changes to the programming were too difficult to resolve and it was simply better to close the pull request altogether.  That is why this commit makes use of the few good things from pull request #12.