Incorporate signal usage
On the secondary controller, you can now use the right stick to control the signal LEDs.  Move the right stick up to signal for cone, and move it down to signal for cube.  Note that, for safety reasons, you will not be able to correct the robot's yaw while using the right stick to signal to the human player.