Remove Ramprate
Often, the wheels would completely change direction and the ramprate would tear the carpet.  For example, if I am going right and I stop and rotate instead, the wheels are spinning in opposite directions for a while as they slow to a complete stop.  Because of this, ramprates have been removed, thus producing version 4.9 of Weswerve.