Add arm enable/disable
Back and Start on the secondary controller can be used to disable the robotic arm completely (BACK) or re-enable it (START).  This is useful, especially if the battery is getting very low, because it can stop wasting so much power on the robotic arm if the robotic arm is not needed.  The robotic arm does require a constant amount of power to be sent to it in order for it to uphold itself.