Delete WPILib-License.md
License is not applicable as the code in this repository is not written by "FIRST and other WPILi contributors".  This code is written by team 5535, the Bionic Bison, and all code rights belong to this team.  As for Weswerve.java, all rights of this file and its usage belong to Wesley McGinn {wesleymcginn1@gmail.com}.