Fix drop function
Previously, the robot was not accurately keeping track of the score because, for the few milliseconds that the score button was held down, this function would be run, thinking that it was scoring hundreds of game pieces per second.  Now, the robot may only count one game piece per three seconds.