Make GetObject Smart Feature FMS-Based
Previously, the robot relied on `counts`, basically taking the number of times the periodic loop had run as a timer in itself.  This was not a good strategy as the robot's thinking speed was not always consistent.  Now, the robot relies on the FMS timer to keep track of timeouts and waiting periods.